requirements.txt
# Core
Django==3.2.14
django-environ==0.9.0
gunicorn==20.1.0
psycopg2==2.9.5
django-anymail[amazon-ses]==10.2
boto3==1.34.58

# For the persistence stores
psycopg2-binary==2.9.3

# Rest apis
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
django-cors-headers==3.13.0
drf-yasg==1.21.5
whitenoise==6.3.0

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1
flake8-isort==4.1.1
coverage==6.4.2
black==22.6.0
pylint-django==2.5.3
pre-commit==2.20.0

# Certificates
cryptography~=37.0.4

django-filter==23.2
sqs-client==0.1.4
